genesis-imran_setup/railtracks-agent/pyproject.toml
[project]
name = "mindpulse-agent"
version = "0.1.0"
dependencies = ["railtracks", "fastapi", "uvicorn", "python-dotenv"]
